storage: share SchemaChange across Postgres and MySQL sources
Move the SchemaChange enum, KeyRef, and the recovery hint into a new
mz-source-schema-change crate that both mz-postgres-util and
mz-mysql-util depend on. Each source keeps its own SchemaChangeError
wrapper naming the affected table, so Postgres still carries and renders
the table oid while MySQL has none. TableRenamed carries the new table's
oid as an Option and renders it only when present.
