+++ Cargo.toml
@@ -96,6 +96,7 @@ members = [
     "src/segment",
     "src/server-core",
     "src/service",
+    "src/source-schema-change",
     "src/sql",
     "src/sql-anonymize",
     "src/sql-lexer",
@@ -222,6 +223,7 @@ default-members = [
     "src/segment",
     "src/server-core",
     "src/service",
+    "src/source-schema-change",
     "src/sql",
     "src/sql-anonymize",
     "src/sql-lexer",
